Refactoring functions and adding comments

+++ RelayTestApp/Plugins/BCClient/Source/BCClientPlugin/Public/BCBlueprintRelayCallProxyBase.h
@@ -5,7 +5,7 @@
 #include "IRelayCallback.h"
 #include "BCBlueprintRelayCallProxyBase.generated.h"
 
-DECLARE_DYNAMIC_MULTICAST_DELEGATE_TwoParams(FBrainCloudRelayCallbackDelegate, uint8, netId, const TArray<uint8>&, data);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_TwoParams(FBrainCloudRelayCallbackDelegate, int, netId, const TArray<uint8>&, data);
 
 UCLASS(MinimalAPI)
 class UBCBlueprintRelayCallProxyBase : public UObject, public IRelayCallback
@@ -20,7 +20,7 @@ class UBCBlueprintRelayCallProxyBase : public UObject, public IRelayCallback
     FBrainCloudRelayCallbackDelegate OnRelayCallback;
 
     //callbacks
-    void relayCallback(uint8 netId,const TArray<uint8>& in_data)
+    void relayCallback(int netId,const TArray<uint8>& in_data)
     {
         OnRelayCallback.Broadcast(netId,in_data);
     }
